App: Break infinite cycle in `GroupExtension::removeObjectsFromDocument`.
The existing code could end up in an infinite loop when removing group
objects from the document using
`GroupExtension::removeObjectsFromDocument`.

The code ends up in `Document::removeObject(const char* sName)` which
will return early in case the object is pending recompute:

```cpp
    if (object->testStatus(ObjectStatus::PendingRecompute)) {
        // ...
FC_MSG("pending remove of " << sName << " after recomputing
document " << getName());
        d->pendingRemove.emplace_back(object);
        return;
    }
```

When this happens, the group size remains the same, since the object has
not been removed yet.

The main loop in `GroupExtension::removeObjectsFromDocument()` was
processing objects as long as the group was not empty:

```cpp
while (Group.getSize() > 0)
```

But this condition would remain true forever. This tracks which object
are pending removal to break out of the cycle.